test(ui): extract shared stubWorker helper to cut duplication (Sonar gate)
SonarCloud's new-code duplication gate failed (13.7% > 3%): the two worker
middleware tests carried an identical CapturedWorker interface + MockWorker
recorder + global-stub block. Hoist that into src/test/workerStub.ts and have
both tests import { CapturedWorker, stubWorker }; behavior and per-file setup()
are unchanged.
